requirements.txt
numpy==1.24.4
pytorch_lightning==2.1.3
scipy~=1.10.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1

scikit-image~=0.21.0
matplotlib~=3.7.2
pillow~=10.0.1